+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [4.6.2] - 2026-07-15
+
 ### Fixes
 
 - *(help)* Say `alias` when there is only one
@@ -4936,7 +4938,8 @@ Minimum version of Rust is now v1.13.0 (Stable)
 * **arg**  allow lifetimes other than 'static in arguments ([9e8c1fb9](https://github.com/clap-rs/clap/commit/9e8c1fb9406f8448873ca58bab07fe905f1551e5))
 
 <!-- next-url -->
-[Unreleased]: https://github.com/clap-rs/clap/compare/v4.6.1...HEAD
+[Unreleased]: https://github.com/clap-rs/clap/compare/v4.6.2...HEAD
+[4.6.2]: https://github.com/clap-rs/clap/compare/v4.6.1...v4.6.2
 [4.6.1]: https://github.com/clap-rs/clap/compare/v4.6.0...v4.6.1
 [4.6.0]: https://github.com/clap-rs/clap/compare/v4.5.61...v4.6.0
 [4.5.61]: https://github.com/clap-rs/clap/compare/v4.5.60...v4.5.61